create scopes users

+++ controller/delete-todos.go
@@ -2,6 +2,7 @@ package controller
 
 import (
 	"database/sql"
+	"go-todo-list/models"
 	"net/http"
 
 	"github.com/labstack/echo"
@@ -10,11 +11,13 @@ import (
 func DeleteTodos(e *echo.Echo, db *sql.DB) {
 
 	e.DELETE("/todos/:id", func(ctx echo.Context) error {
+		user := ctx.Get("USER").(models.AuthClaimJWT)
 		id := ctx.Param("id")
 
 		_, err := db.Exec(
-			"DELETE FROM todos WHERE id = ?",
+			"DELETE FROM todos WHERE id = ? AND user_id = ?",
 			id,
+			user.UserId,
 		)
 		if err != nil {
 			return ctx.String(http.StatusInternalServerError, err.Error())
